created class NumberRules which contains logic for filtering numbers
